feat: SettingViewModuel 생성및 파일 이동
