+++ checky/checky/View/Setting/Event/EventSettingViewModel.swift
@@ -7,10 +7,13 @@
 
 import Foundation
 import EventKit
+import SwiftUI
+import Combine
 
 class EventSettingViewModel: ObservableObject {
   let eventManager: EventManager
   @Published var categories: [EKCalendar]
+  @Published var color: Color = .white
 
   init(
     eventManager: EventManager
