--- checky/checky/View/Setting/SettingViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SettingViewModel.swift
-//  checky
-//
-//  Created by song on 2022/11/01.
-//
-
-import Foundation
