+++ checky/checky/View/Setting/ViewModuel/SelectorView/ColorView.swift
