+++ checky/checky/View/Setting/ViewModuel/SelectorView/EmojiView.swift
